% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassImpacts.R
\name{ClassImpacts}
\alias{ClassImpacts}
\alias{ClassImpacts.Context}
\title{Classification impacts class builder}
\usage{
ClassImpacts(
  context,
  region,
  incursion,
  impact_layers,
  impact_classes,
  combine_function = c("max", "none"),
  mgmt_costs = NULL,
  ...
)

\method{ClassImpacts}{Context}(
  context,
  region,
  incursion,
  impact_layers,
  impact_classes,
  combine_function = c("max", "none"),
  mgmt_costs = NULL,
  ...
)
}
\arguments{
\item{context}{A \code{Context} or inherited class object representing the
context of the impact analysis.}

\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the impact analysis.}

\item{incursion}{An \code{Incursion} or inherited class object representing
the spatial distribution of an actual invasive species presence or
density, or incursion likelihoods across the \code{region}.}

\item{impact_layers}{A named list of binary (0,1) spatial layers
(\code{terra::SpatRaster} or \code{raster::RasterLayer}) or binary vectors
of locations (consistent with the spatial \code{region}), for indicating
the spatial distribution of each named aspect (mechanism, service, sector,
asset type, etc.) specified via the impact scope in the \code{context}.
Unknown layers/values for aspects may be omitted.}

\item{impact_classes}{A vector of class (ranking or category) values
(consistent with the impact measures in the \code{context}) for each named
aspect (mechanism, service, sector, asset type, etc.) specified via the
impact scope in the \code{context}.}

\item{combine_function}{The function used to combine impact layers across
aspects of the environment, society, and/or economy. Either \code{"max"},
(for maximum ordinal class), or a user defined function having the form
\code{"function(aspect_locations)"}, where \code{aspect_locations} is
a list of vectors of the class at each \code{region} location for each
\code{context} impact scope aspect, which is passed to the function.
The function should return a single vector of classes for each location.
Set to \code{"none"} when combining impacts is not applicable.}

\item{mgmt_costs}{Optional spatial layer (\code{terra::SpatRaster} or
\code{raster::RasterLayer}) or vector of management costs at each
location specified by the \code{region}, measured in the unit specified
in the \code{context}. Default is \code{NULL}.}

\item{...}{Additional parameters.}
}
\value{
A \code{ClassImpacts} class object (list) containing functions
for classifying and combining invasive species (likely) incursion impacts,
and management costs (optional):
\describe{
\item{\code{incursion_impacts()}}{Classify (likely) incursion impacts
for each aspect of the environment, society, and/or economy.}
\item{\code{combined_impacts()}}{Combine (likely) incursion impacts
across aspects of the environment, society, and/or economy, to produce
an overall impact class at each location.}
\item{\code{incursion_mgmt_costs()}}{Calculate (likely) incursion
management costs at each location (when specified).}
}
}
\description{
Builds a class to represent classification impact analysis functionality for
ranking, categorising, and combining spatially-explicit class-based impacts
of invasive species incursions across various aspects of the environment,
society, and/or economy.
}
