% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Context.R
\name{Context}
\alias{Context}
\alias{Context.default}
\title{Context class builder}
\usage{
Context(
  species_name,
  species_type = c("pest", "weed", "disease"),
  impact_types = c("ecological", "social", "economic"),
  impact_scope,
  valuation_type = c("monetary", "non-monetary", "ranking", "categorical"),
  impact_measures,
  ...
)

\method{Context}{default}(
  species_name,
  species_type = c("pest", "weed", "disease"),
  impact_types = c("ecological", "social", "economic"),
  impact_scope,
  valuation_type = c("monetary", "non-monetary", "ranking", "categorical"),
  impact_measures,
  ...
)
}
\arguments{
\item{species_name}{Invasive species, genus, or functional group name.}

\item{species_type}{Type of invasive species. One of \code{"pest"},
\code{"weed"}, or \code{"disease"}.}

\item{impact_types}{The types of impacts involved. One or more of
\code{"ecological"}, \code{"social"}, and/or \code{"economic"}.}

\item{impact_scope}{A vector of aspects of the impact types analysed, such
as ecological mechanisms, ecosystem services, social aspects, economic
sectors and/or asset types.}

\item{valuation_type}{The type of valuation used to measure impacts. One of
\code{"monetary"}, \code{"non-monetary"} (quantities), \code{"ranking"},
or \code{"categorical"}.}

\item{impact_measures}{A vector of one or more measures used to quantify or
classify each impact aspect, consistent with the valuation type. Monetary
or non-monetary (quantitative) measures should specify the unit used (e.g.
"$"). Ranking or categorical measures should specify a list of
values/categories in ascending order (when applicable), such as GISS
(0, 1, 2, 3, 4, 5) ranking, or EICAT/SEICAT ("MC", "MN", "MO", "MR", "MV")
categories.}

\item{...}{Additional parameters.}
}
\value{
A \code{Context} class object (list) containing functions for
accessing attributes:
\describe{
\item{\code{get_species_name()}}{Get the invasive species, genus, or
functional group name.}
\item{\code{get_species_type()}}{Get the type of invasive species:
"pest", "weed", or "disease".}
\item{\code{get_impact_types()}}{Get the types of impacts analysed:
"ecological", "social", and/or "economic".}
\item{\code{get_impact_scope()}}{Get the aspects of impacts analysed.}
\item{\code{get_valuation_type()}}{Get the type of valuation used to
measure impacts: "monetary", "non-monetary", "ranking",
"categorical".}
\item{\code{get_impact_measures()}}{Get the measures used to quantify or
classify each impact aspect.}
}
}
\description{
Builds a class to represent the context of a bio-security impact analysis,
including: information about the invasive pest, weed or disease species,
genus or functional group involved; the types of impacts involved
(ecological/environmental, social/human-welfare, and/or economic/market);
the scope of impacts, such as aspects of environment (e.g. ecological
mechanisms), social aspects (e.g. culture, health), and/or economic sectors;
the valuation or measures used, such as monetary values, other quantified
values (e.g. national significance, distinctiveness), ranking (e.g. GISS) or
classification (e.g. EICAT, SEICAT).
}
