% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpactAnalysis.R
\name{ImpactAnalysis}
\alias{ImpactAnalysis}
\alias{ImpactAnalysis.Context}
\title{Impact analysis base class builder}
\usage{
ImpactAnalysis(
  context,
  region,
  incursion,
  impact_layers,
  combine_function = c("sum", "max"),
  mgmt_costs = NULL,
  subclass = character(),
  ...
)

\method{ImpactAnalysis}{Context}(
  context,
  region,
  incursion,
  impact_layers,
  combine_function = c("sum", "max"),
  mgmt_costs = NULL,
  subclass = character(),
  ...
)
}
\arguments{
\item{context}{A \code{Context} or inherited class object representing the
context of the impact analysis.}

\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the impact analysis.}

\item{incursion}{An \code{Incursion} or inherited class object representing
the spatial distribution of an actual invasive species presence or
density, or incursion likelihoods across the \code{region}.}

\item{impact_layers}{A named list of spatial layers
(\code{terra::SpatRaster} or \code{raster::RasterLayer}) or vectors of
location values (consistent with the spatial \code{region}), for each
named aspect (mechanism, service, sector, asset type, etc.) specified via
the impact scope in the \code{context}. Unknown layers/values for aspects
may be omitted.}

\item{combine_function}{The function used to combine impact layers across
aspects of the environment, society, and/or economy. Either \code{"sum"},
\code{"max"} (maximum), or a user defined function having the form
\code{"function(aspect_locations)"}, where \code{aspect_locations} is
a list of vectors of values at each \code{region} location for each
\code{context} impact scope aspect, which is passed to the function.
The function should return a single vector of values for each location.}

\item{mgmt_costs}{Optional spatial layer (\code{terra::SpatRaster} or
\code{raster::RasterLayer}) or vector of management costs at each
location specified by the \code{region}, measured in the unit specified
in the \code{context}. Default is \code{NULL}.}

\item{...}{Additional parameters.}
}
\value{
An \code{ImpactAnalysis} class object (list) containing functions
for calculating invasive species (likely) incursion impacts and management
costs (optional):
\describe{
\item{\code{incursion_impacts()}}{Calculate (likely) incursion impacts
for each aspect of the environment, society, and/or economy.}
\item{\code{combined_impacts()}}{Combine (likely) incursion impacts
across aspects of the environment, society, and/or economy, to produce
an overall impact at each location.}
\item{\code{incursion_mgmt_costs()}}{Calculate (likely) incursion
management costs at each location (when specified).}
}
}
\description{
Builds a base class to represent impact analysis functionality for
calculating or classifying, and combining, spatially-explicit impacts of
invasive species incursions across various aspects of the environment,
society, and/or economy.
}
