% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incursion.R
\name{Incursion}
\alias{Incursion}
\alias{Incursion.Raster}
\alias{Incursion.SpatRaster}
\alias{Incursion.default}
\title{Incursion class builder}
\usage{
Incursion(
  x,
  region = NULL,
  type = c("presence", "density", "prob"),
  multiplier = 1,
  threshold = 0,
  ...
)

\method{Incursion}{Raster}(x, ...)

\method{Incursion}{SpatRaster}(x, region = NULL, ...)

\method{Incursion}{default}(
  x,
  region = NULL,
  type = c("presence", "density", "prob"),
  multiplier = 1,
  threshold = 0,
  ...
)
}
\arguments{
\item{x}{A \code{terra::SpatRaster}, \code{raster::RasterLayer}, or numeric
vector defining the incursion presence, density, or probabilities at each
spatial location.}

\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for impact analysis.}

\item{type}{One of \code{"presence"}, \code{"density"}, or \code{"prob"} to
indicate if the values in \code{x} represent incursion presence, density,
or probabilities respectively.}

\item{multiplier}{Numeric multiplier to transform/scale incursion values for
calculating impacts (especially when type is \code{"density"}). Default is
\code{1}.}

\item{threshold}{Numeric threshold (>= value) for signifying sufficient
incursion density or probability to contribute to impact calculations
(after the application of the \code{multiplier}). Default is \code{0}.}

\item{...}{Additional parameters.}
}
\value{
An \code{Incursions} class object (list) containing functions for
accessing incursion type and transformed incursion values:
\describe{
\item{\code{get_type()}}{Get the incursion type.}
\item{\code{get_impact_incursion()}}{Get the transformed incursion (0-1)
values for calculating impacts.}
}
}
\description{
Builds a class to represent a spatially-explicit incursion (layer) of actual
invasive species presence or density, or incursion likelihoods across a
defined region for impact analysis, and transforms the incursion values for
impact calculations.
}
