% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region.R
\name{Region}
\alias{Region}
\alias{Region.Raster}
\alias{Region.SpatRaster}
\alias{Region.matrix}
\alias{Region.data.frame}
\alias{Region.NULL}
\title{Region class builder}
\usage{
Region(x, ...)

\method{Region}{Raster}(x, ...)

\method{Region}{SpatRaster}(x, ...)

\method{Region}{matrix}(x, ...)

\method{Region}{data.frame}(x, ...)

\method{Region}{`NULL`}(...)
}
\arguments{
\item{x}{A \code{terra::SpatRaster} or \code{raster::RasterLayer} object
representing a grid-based spatial region (template). The spatial division
of locations, patches, or sub-regions may also be defined via a data frame
(or matrix) of location coordinates in longitude and latitude (WGS84) with
explicitly named columns "lon" and "lat". If missing, the region is
assumed to be a single aspatial patch.}

\item{...}{Additional parameters.}
}
\value{
A \code{Region} class object (list) containing functions for
accessing attributes, checking compatibility of objects with the
region, and to maintain and calculate spatial data:
\describe{
\item{\code{get_type()}}{Get the type of representation: spatial "grid"
(raster cells), "patch" (coordinates), or "single" (aspatial).}
\item{\code{get_locations()}}{Get the number of locations (cells or
patches) that are included in the simulation.}
\item{\code{is_compatible(y)}}{Check the compatibility of object
\code{y} with the region defined by \code{x}.}
\item{\code{get_template(empty = FALSE)}}{Get the spatial template when
the \code{type} is "grid", with either zeros in non-NA locations
(default), or with no values when \code{empty = TRUE}.}
\item{\code{get_rast(values)}}{Get spatial \code{terra::SpatRaster} set
with a single value or a vector of values matching the number of
locations when \code{type} is "grid".}
\item{\code{get_indices()}}{Get cell indices of grid or patch locations
that are included in the simulation when the \code{type} is "grid".}
\item{\code{get_res()}}{Get the spatial cell resolution (in m) of the
region when the \code{type} is "grid".}
\item{\code{is_included(indices)}}{Check if cell \code{indices} of grid
locations are included (non-NA cells) in the simulation when the
\code{type} is "grid", and return a logical vector indicating the
inclusion of each index.}
\item{\code{get_data()}}{Get the data frame specifying the division when
\code{type} is "patch" or "other".}
\item{\code{get_coords(extra_cols = FALSE)}}{Get a data frame of patch
location coordinates when \code{type} is "patch", as well as optional
extra named columns from the original location data.}
\item{\code{get_feat()}}{Get spatial \code{terra::SpatVector} features
or points when \code{type} is "grid" or "patch".}
}
}
\description{
Builds a class to represent a spatial region for an impact analysis,
defined via a raster layer with active (non-NA) cells, or via a data
frame specifying locations, patches, or sub-regions.
}
