% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValueImpacts.R
\name{ValueImpacts}
\alias{ValueImpacts}
\alias{ValueImpacts.Context}
\title{Value impacts class builder}
\usage{
ValueImpacts(
  context,
  region,
  incursion,
  impact_layers,
  loss_rates,
  mgmt_costs = NULL,
  combine_function = "sum",
  ...
)

\method{ValueImpacts}{Context}(
  context,
  region,
  incursion,
  impact_layers,
  loss_rates,
  mgmt_costs = NULL,
  combine_function = "sum",
  ...
)
}
\arguments{
\item{context}{A \code{Context} or inherited class object representing the
context of the impact analysis.}

\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the impact analysis.}

\item{incursion}{An \code{Incursion} or inherited class object representing
the spatial distribution of an actual invasive species presence or
density, or incursion likelihoods across the \code{region}.}

\item{impact_layers}{A named list of spatial layers
(\code{terra::SpatRaster} or \code{raster::RasterLayer}) or vectors of
location values (consistent with the spatial \code{region}), for each
named aspect (mechanism, service, sector, asset type, etc.) specified via
the impact scope in the \code{context}.}

\item{loss_rates}{A vector of value loss rates for each named aspect
(mechanism, service, sector, asset type, etc.) specified via the impact
scope in the \code{context}.}

\item{mgmt_costs}{A spatial layer (\code{terra::SpatRaster} or
\code{raster::RasterLayer}) or vector of management costs at each
location (specified by the \code{region}). Only used when the
\code{context} valuation type is \code{"monetary"}. Default is
\code{NULL}.}

\item{combine_function}{The function used to combine impact layers across
aspects of the environment, society, and/or economy. Either \code{"sum"}
or a user defined function having the form
\code{"function(aspect_locations)"}, where \code{aspect_locations} is
a list of vectors of values at each \code{region} location for each
\code{context} impact scope aspect, which is passed to the function.
The function should return a single vector of values for each location.}

\item{...}{Additional parameters.}
}
\value{
An \code{ValueImpacts} class object (list) containing functions
for calculating invasive species (likely) incursion impacts and management
costs (\code{context} valuation \code{"monetary"} type only):
\describe{
\item{\code{incursion_impacts()}}{Calculate (likely) incursion impacts
(damages or losses) for each aspect of the environment, society,
and/or economy.}
\item{\code{combined_impacts()}}{Combine (likely) incursion impacts
across aspects of the environment, society, and/or economy, to produce
an overall impact (damage or loss).}
\item{\code{incursion_mgmt_costs()}}{Calculate (likely) incursion
management costs (valuation \code{"monetary"} type only).}
\item{\code{total_costs()}}{Calculate (likely) total incursion
(damages/losses plus management) costs (valuation \code{"monetary"}
type only).}
}
}
\description{
Builds a class to represent quantitative impact analysis functionality for
calculating and combining spatially-explicit value-based impacts of invasive
species incursions across various aspects of the environment, society,
and/or economy.
}
