% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassImpacts.R
\name{ClassImpacts}
\alias{ClassImpacts}
\alias{ClassImpacts.Context}
\title{Classification impacts class builder}
\usage{
ClassImpacts(
  context,
  region,
  incursion,
  impact_layers,
  impact_classes,
  combine_function = c("max", "none"),
  mgmt_costs = NULL,
  ...
)

\method{ClassImpacts}{Context}(
  context,
  region,
  incursion,
  impact_layers,
  impact_classes,
  combine_function = c("max", "none"),
  mgmt_costs = NULL,
  ...
)
}
\arguments{
\item{context}{A \code{Context} or inherited class object representing the
context of the impact analysis.}

\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the impact analysis.}

\item{incursion}{An \code{Incursion} or inherited class object representing
the spatial distribution of an actual invasive species presence or
density, or incursion likelihoods across the \code{region}.}

\item{impact_layers}{A named list of binary (0,1) spatial layers
(\code{terra::SpatRaster} or \code{raster::RasterLayer}) or binary vectors
of locations (consistent with the spatial \code{region}), for indicating
the spatial distribution of each named aspect (mechanism, service, sector,
asset type, etc.) specified via the impact scope in the \code{context}.
Unknown layers/values for aspects may be omitted.}

\item{impact_classes}{A vector of class (ranking or category) values
(consistent with the impact measures in the \code{context}) for each named
aspect (mechanism, service, sector, asset type, etc.) specified via the
impact scope in the \code{context}.}

\item{combine_function}{The function used to combine impact layers across
aspects of the environment, society, and/or economy. Either \code{"max"},
(for maximum ordinal class), or a user defined function having the form
\code{"function(aspect_locations)"}, where \code{aspect_locations} is
a list of vectors of the class at each \code{region} location for each
\code{context} impact scope aspect, which is passed to the function.
The function should return a single vector of classes for each location.
Set to \code{"none"} when combining impacts is not applicable.}

\item{mgmt_costs}{Optional spatial layer (\code{terra::SpatRaster} or
\code{raster::RasterLayer}) or vector of management costs at each
location specified by the \code{region}, measured in the unit specified
in the \code{context}. Default is \code{NULL}.}

\item{...}{Additional parameters.}
}
\value{
A \code{ClassImpacts} class object (list) containing functions
for classifying and combining invasive species (likely) incursion impacts,
and management costs (optional):
\describe{
\item{\code{get_context()}}{Get context object.}
\item{\code{get_incursion()}}{Get incursion object.}
\item{\code{get_id()}}{Get the impacts numeric identifier.}
\item{\code{set_id(id)}}{Set the impacts numeric identifier.}
\item{\code{incursion_impacts(raw = FALSE)}}{Classify (likely)
incursion impacts for each aspect of the environment, society, and/or
economy. Returns results consistent with region, or vectors when
\code{raw = TRUE}.}
\item{\code{combined_impacts(raw = FALSE)}}{Combine (likely) incursion
impacts across aspects of the environment, society, and/or economy, to
produce an overall impact at each location. Returns result consistent
with region, or a vector when \code{raw = TRUE}.}
\item{\code{incursion_mgmt_costs()}}{Calculate (likely) incursion
management costs at each location (when specified).}
\item{\code{save_analysis(...)}}{Save the impact analysis as a
collection of raster TIF and/or comma-separated value (CSV) files,
appropriate for the \code{region} type, including the individual and
combined incursion impacts, and incursion management costs (when
specified).\code{Terra} raster write options may be passed to the
function for saving grid-based analysis outputs.}
}
}
\description{
Builds a class to represent classification impact analysis functionality for
ranking, categorising, and combining spatially-explicit class-based impacts
of invasive species incursions across various aspects of the environment,
society, and/or economy.
}
\references{
ABARES 2021, The National Priority List of Exotic Environmental Pests,
Weeds and Diseases: Information Paper (Version 2.0), ABARES report to
client prepared for the Chief Environmental Biosecurity Officer,
Department of Agriculture, Water and the Environment, Canberra, ACT. CC
BY 4.0.
\url{https://www.agriculture.gov.au/sites/default/files/documents/eepl-information-paper.pdf}

Bacher, S., Blackburn, T. M., Essl, F., Genovesi, P., Heikkilä, J.,
Jeschke, J. M., Jones, G., Keller, R., Kenis, M., Kueffer, C.,
Martinou, A. F., Nentwig, W., Pergl, J., Pyšek, P., Rabitsch, W.,
Richardson, D. M., Roy, H. E., Saul, W-C., Scalera, R., Vilà, M.,
Wilson, J. R. U., Kumschick, S. (2018). Socio-economic impact
classification of alien taxa (SEICAT).
\emph{Methods in Ecology and Evolution}, 9(1), 159–168.
\doi{10.1111/2041-210X.12844}

Blackburn, T. M., Essl, F., Evans, T., Hulme, P. E., Jeschke, J. M.,
Kühn, I., Kumschick, S., Marková, Z., Mrugała, A., Nentwig, W., Pergl, J.,
Pyšek, P., Rabitsch, W., Ricciardi, A., Richardson, D. M., Sendek, A.,
Vilà, M., Wilson, J. R. U., Winter, M., Genovesi, P., & Bacher, S. (2014).
A Unified Classification of Alien Species Based on the Magnitude of their
Environmental Impacts. \emph{PLoS Biology}, 12(5).
\doi{10.1371/journal.pbio.1001850}

Ireland, K. B., van Klinken, R., Cook, D. C., Jamieson, L., Hulme, P. E.,
Worner, S., Rodoni, B., Teulon, D., Crampton, K. A., Hodda, M., Paini, D.,
Logan, D., Tyson, J. L., Brockerhoff, E. G., Fletcher, J. D.,
Christopher, M., Ludowici, V. A., & Bulman, L. (2020). Plant Pest Impact
Metric System (PPIMS): Framework and guidelines for a common set of
metrics to classify and prioritise plant pests. \emph{Crop Protection},
128. \doi{10.1016/j.cropro.2019.105003}

IUCN (2020). IUCN EICAT Categories and Criteria. The Environmental Impact
Classification for Alien Taxa (EICAT) First edition.
\emph{IUCN, Gland, Switzerland and Cambridge, UK}. IUCN.
\doi{10.2305/IUCN.CH.2020.05.en}

Nentwig, W., Kühnel, E., & Bacher, S. (2010). A Generic Impact-Scoring
System Applied to Alien Mammals in Europe. \emph{Conservation Biology},
24(1), 302–311. \doi{10.1111/j.1523-1739.2009.01289.x}
}
