% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValueImpacts.R
\name{ValueImpacts}
\alias{ValueImpacts}
\alias{ValueImpacts.Context}
\title{Value impacts class builder}
\usage{
ValueImpacts(
  context,
  region,
  incursion,
  impact_layers,
  loss_rates,
  combine_function = c("sum", "none"),
  mgmt_costs = NULL,
  ...
)

\method{ValueImpacts}{Context}(
  context,
  region,
  incursion,
  impact_layers,
  loss_rates,
  combine_function = c("sum", "none"),
  mgmt_costs = NULL,
  ...
)
}
\arguments{
\item{context}{A \code{Context} or inherited class object representing the
context of the impact analysis.}

\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the impact analysis.}

\item{incursion}{An \code{Incursion} or inherited class object representing
the spatial distribution of an actual invasive species presence or
density, or incursion likelihoods across the \code{region}.}

\item{impact_layers}{A named list of spatial layers
(\code{terra::SpatRaster} or \code{raster::RasterLayer}) or vectors of
location values (consistent with the spatial \code{region}), for each
named aspect (mechanism, service, sector, asset type, etc.) specified via
the impact scope in the \code{context}.}

\item{loss_rates}{A vector of value loss rates for each named aspect
(mechanism, service, sector, asset type, etc.) specified via the impact
scope in the \code{context}.}

\item{combine_function}{The function used to combine impact layers across
aspects of the environment, society, and/or economy. Either \code{"sum"}
or a user defined function having the form
\code{"function(aspect_locations)"}, where \code{aspect_locations} is
a list of vectors of values at each \code{region} location for each
\code{context} impact scope aspect, which is passed to the function.
The function should return a single vector of values for each location.
Set to \code{"none"} when combining impacts is not applicable.}

\item{mgmt_costs}{Optional spatial layer (\code{terra::SpatRaster} or
\code{raster::RasterLayer}) or vector of management costs at each
location specified by the \code{region}, measured in the unit specified
in the \code{context}. Default is \code{NULL}.}

\item{...}{Additional parameters.}
}
\value{
A \code{ValueImpacts} class object (list) containing functions
for calculating and combining invasive species (likely) incursion impacts,
management costs (optional), and total costs (when applicable):
\describe{
\item{\code{incursion_impacts()}}{Calculate (likely) incursion impacts
(damages or losses) for each aspect of the environment, society,
and/or economy.}
\item{\code{combined_impacts()}}{Combine (likely) incursion impacts
across aspects of the environment, society, and/or economy, to produce
an overall impact (damage or loss) at each location.}
\item{\code{incursion_mgmt_costs()}}{Calculate (likely) incursion
management costs (when specified) at each location.}
\item{\code{total_costs()}}{Calculate (likely) total incursion
(damages/losses plus management) costs at each location (when
\code{context} \code{mgmt_cost_unit} matches \code{impact_measures}).}
}
}
\description{
Builds a class to represent quantitative impact analysis functionality for
calculating and combining spatially-explicit value-based impacts of invasive
species incursions across various aspects of the environment, society,
and/or economy.
}
