% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValueImpacts.R
\name{ValueImpacts}
\alias{ValueImpacts}
\alias{ValueImpacts.Context}
\title{Value impacts class builder}
\usage{
ValueImpacts(
  context,
  region,
  incursion,
  impact_layers,
  loss_rates,
  combine_function = c("sum", "none"),
  mgmt_costs = NULL,
  ...
)

\method{ValueImpacts}{Context}(
  context,
  region,
  incursion,
  impact_layers,
  loss_rates,
  combine_function = c("sum", "none"),
  mgmt_costs = NULL,
  ...
)
}
\arguments{
\item{context}{A \code{Context} or inherited class object representing the
context of the impact analysis.}

\item{region}{A \code{Region} or inherited class object representing the
spatial region (template) for the impact analysis.}

\item{incursion}{An \code{Incursion} or inherited class object representing
the spatial distribution of an actual invasive species presence or
density, or incursion likelihoods across the \code{region}.}

\item{impact_layers}{A named list of spatial layers
(\code{terra::SpatRaster} or \code{raster::RasterLayer}) or vectors of
location values (consistent with the spatial \code{region}), for each
named aspect (mechanism, service, sector, asset type, etc.) specified via
the impact scope in the \code{context}.}

\item{loss_rates}{A vector of value loss rates for each named aspect
(mechanism, service, sector, asset type, etc.) specified via the impact
scope in the \code{context}.}

\item{combine_function}{The function used to combine impact layers across
aspects of the environment, society, and/or economy. Either \code{"sum"}
or a user defined function having the form
\code{"function(aspect_locations)"}, where \code{aspect_locations} is
a list of vectors of values at each \code{region} location for each
\code{context} impact scope aspect, which is passed to the function.
The function should return a single vector of values for each location.
Set to \code{"none"} when combining impacts is not applicable.}

\item{mgmt_costs}{Optional spatial layer (\code{terra::SpatRaster} or
\code{raster::RasterLayer}) or vector of management costs at each
location specified by the \code{region}, measured in the unit specified
in the \code{context}. Default is \code{NULL}.}

\item{...}{Additional parameters.}
}
\value{
A \code{ValueImpacts} class object (list) containing functions
for calculating and combining invasive species (likely) incursion impacts,
management costs (optional), and total costs (when applicable):
\describe{
\item{\code{get_context()}}{Get context object.}
\item{\code{get_incursion()}}{Get incursion object.}
\item{\code{incursion_impacts(raw = FALSE)}}{Calculate (likely)
incursion impacts (damages or losses) for each aspect of the
environment, society, and/or economy. Returns results consistent with
region, or vectors when \code{raw = TRUE}.}
\item{\code{combined_impacts(raw = FALSE)}}{Combine (likely) incursion
impacts across aspects of the environment, society, and/or economy, to
produce an overall impact (damage or loss) at each location. Returns
result consistent with region, or a vector when \code{raw = TRUE}.}
\item{\code{incursion_mgmt_costs()}}{Calculate (likely) incursion
management costs (when specified) at each location.}
\item{\code{total_costs()}}{Calculate (likely) total incursion
(damages/losses plus management) costs at each location (when
\code{context} \code{mgmt_cost_unit} matches \code{impact_measures}).}
\item{\code{save_analysis(...)}}{Save the impact analysis as a
collection of raster TIF and/or comma-separated value (CSV) files,
appropriate for the \code{region} type, including the individual and
combined incursion impacts, and incursion management and total costs
(when specified).\code{Terra} raster write options may be passed to
the function for saving grid-based analysis outputs.}
}
}
\description{
Builds a class to represent quantitative impact analysis functionality for
calculating and combining spatially-explicit value-based impacts of invasive
species incursions across various aspects of the environment, society,
and/or economy.
}
\references{
Arponen, A., Heikkinen, R. K., Thomas, C. D., & Moilanen, A. (2005). The
Value of Biodiversity in Reserve Selection: Representation, Species
Weighting, and Benefit Functions. \emph{Conservation Biology}, 19(6),
2009–2014. \url{https://www.jstor.org/stable/3591223}

Costanza, R., D'Arge, R., De Groot, R., Farber, S., Grasso, M.,
Hannon, B., Limburg, K., Naeem, S., O'Neill, R. V., & Paruelo, J. (1998).
The value of ecosystem services: putting the issues in perspective.
\emph{Ecological Economics}, 25(E 1), 67–72.
\doi{10.1016/S0921-8009(98)00019-6}

Department of Agriculture, Fisheries and Forestry (DAFF) (2021). National
Environmental Biosecurity Response Agreement 2.0.
\url{https://www.agriculture.gov.au/biosecuritytrade/policy/emergency/nebra}

Dodd, A. J., Stoeckl, N., Baumgartner, J. B. & Kompas, T. F. (2020). Key
Result Summary: Valuing Australia's Biosecurity System. Tech. Rep. 170713,
\emph{Centre of Excellence for Biosecurity Risk Analysis (CEBRA),}
\emph{The University of Melbourne, Melbourne}.
\url{https://cebra.unimelb.edu.au/__data/assets/pdf_file/0020/3535013/CEBRA_Value_Docs_KeyResultSummary_v0.6_Endorsed.pdf}

MacLeod, A., Head, J., & Gaunt, A. (2004). An assessment of the potential
economic impact of Thrips palmi on horticulture in England and the
significance of a successful eradication campaign. \emph{Crop Protection},
23(7), 610–601. \doi{10.1016/j.cropro.2003.11.010}

O'Loughlin, L. S., Gooden, B., Barney, J. N., & Lindenmayer, D. B. (2019).
Surrogacy in invasion research and management : inferring "impact" from
"invasiveness". \emph{Frontiers in Ecology and the Environment}, 17(8),
464–473. \doi{10.1002/fee.2097}

Soliman, T., Mourits, M. C. M., Oude Lansink, A. G. J. M., &
van der Werf, W. (2010). Economic impact assessment in pest risk analysis.
\emph{Crop Protection}, 29(6), 517–524. \doi{10.1016/j.cropro.2009.12.014}

Soliman, T., Mourits, M. C. M., Oude Lansink, A. G. J. M., &
van der Werf, W. (2015). Quantitative economic impact assessment of
invasive plant pests: What does it require and when is it worth the
effort? \emph{Crop Protection}, 69, 9–17.
\doi{10.1016/j.cropro.2014.11.011}

Stoeckl, N., Dodd, A., & Kompas, T. (2020). Values and vulnerabilities:
what assets are protected by Australia's national biosecurity system and
thus at risk of incursion? \emph{Centre of Excellence for Biosecurity}
\emph{Risk Analysis (CEBRA), The University of Melbourne, Melbourne}.
\url{https://cebra.unimelb.edu.au/__data/assets/pdf_file/0011/3898946/CEBRA_Value_Docs_ValuesAndVulnerabilities_v0.4-PostSAC.pdf}

Weitzman, M. L. (1998). The Noah's Ark Problem. \emph{Econometrica},
66(6), 1279–1298. \doi{10.2307/2999617}

Welsh, M. J., Brockerhoff, E. G., Kean, J. M., Phillips, C.,
Stringer, L. D., Vereijssen, J., Turner, J. A., Epanchin-Niell, R. S.,
Monge, J. J., Soliman, T., Robinson, A. P., Kompas, T., Liebhold, A. M.,
& Ormsby, M. (2021). Approaches for estimating benefits and costs of
interventions in plant biosecurity across invasion phases.
\emph{Ecological Applications}, 31(5). \doi{10.1002/eap.2319}
}
